refactor import paths for NuclearDataManager and nuclearParameters
